Use get_n_service_node's poll_block_hash for swam updates
This lets lokid respond with a "unchanged" result when the block hasn't
change which saves a lot of effort & data: currently storage server
pumps around 400kB/s of data from lokid rpc.  The poll_block_hash (added
in lokid 6) lets lokid just not generate or send the redundant data in
the first place.