Update to lmq 1.0.5 & fix compilation
Updates to 1.0.5 with a fix that can cause message reception delays of
up to 250ms.

1.0.5 also broke SS's forward declaration of `string_view` (because it
grew the capability of handling char types other than `char`); the
underlying simple_string_view class is an implementation detail that
isn't meant to be used (and may not exist at all: lokimq uses
std::string_view instead if the compiler provides it).