Update for LMQ 1.1.0 SN list compatibility
Assigning a connection's SN status at connection time is not working
well: it results in excess reconnections and test failures, since
currently the behaviour goes something like this:

- SS has many other SSes connected to it.
- SS restarts
- Other SSes see the closed connection and reconnect
- Many of these reconnections are fast enough that the SS doesn't have
  the SS list loaded yet, and so they all get service node status set to
  false.
- One of the SSes makes a request (either onion routing something, or
  doing a reachability test).
- SS says "BYE" to tell the remote to disconnect.
- Remote disconnects, and the request fails.
- Remote tests again sometime later and reconnects, this time
  gets noticed as a SN and allowed.

There's a related overhang issue where a connected SN keeps its SN
status on the connection as long as it keeps the connection open even if
it is no longer an active SN.

I've changed LokiMQ to make this work better by doing SN status checking
at the time the command is invoked.  This required some API-incompatible
restructuring, however: it isn't feasible, for performance reasons, to
invoke the Allow callback on each request, so this means LMQ needs to
maintain its own list of active service nodes, and SS needs to tell LMQ
whenever the list of service nodes changes.

This commit updates LokiMQ and implements the updating logic in SS.