Remove headers from cmake add_library
Headers aren't supposed to be listed in `add_library` calls and are an
unfortunately common cmake anti-pattern.  (cmake does *not* need headers
listed to know how to check that things need rebuilding when a header
changes, which seems to be the reason people think they have to include
them).

Apparently this antipattern emerged partly because of buggy behaviour in
MSVC pre-2017 that didn't understand how to find headers when loading a
CMake project.