Make --version print without initializing logger
Currently --version is sort of awkward: it doesn't print anything at all
if you run it with `--log-level warning` (or if you have that in a
config file), and it prints the version info *after* initializing (and
printing various info about) the logger, which makes the output kind of
messy.

This simplifies it to print the version (and exit) much earlier in the
startup.

(This also makes a minor tweak to the current `print_version` from
`static` to `inline` so that the linker can throw away the duplicates
rather than keeping duplicates for each compilation unit that includes
version.h.)