Replace embedded nlohmann with updated submodule
The embedded nlohmann (3.6.1) is fairly old and has bugs that causes
compilation problems under gcc-10.  Switch to a submodule (as with most
of the other vendored libs) and updated to the latest version.

Also switch it to multi-header include mode so that we can use the tiny
json_fwd header in headers rather than the full include.