Use lokimq for hex/b64/b32z conversion
Removes a bunch of copied code (for base32z and hex) and boost::detail
code calls (for base64).

Also includes some additional validity checks on base64 decoding that
didn't properly validate that the input is actually base64 or of the
correct length, then used an unsafe memory copy from the
unchecked-length decoded string.