Remove cmake cruft
- loki_add_subdirectory was a unnecessary wrapper around
add_subdirectory: it was an attempt to make an idempotent version of
add_subdirectory, but that isn't needed at all and just adds cruft: the
top-level CMakeLists.txt already includes all the subdirectories so we
can just trust it.

- removed incorrect subdirectory "project()" definitions.

- crypto/CMakeLists.txt pointlessly listed all headers in the source
list.

- set c++ standard in the top-level makefile instead of on each target
since we intentionally want it everywhere.

- Linking directly to pthread/dl with conditional OS checks was wrong;
fix it to be proper cmake (linking to Threads::Threads and
${CMAKE_DL_LIBS}).

- Various cmake files erroneously listed their src directories in their
include paths.

- Made various library linkages PRIVATE instead of PUBLIC where a
transient dependency to dependent targets does not make sense.