Remove storage-server blockchain testing
This is a mis-feature in the wrong place that doesn't protect against
anything malicious.

There is one comment (in oxen-core) that suggests the purpose here was
to be complex enough that the result couldn't be proxied to some other
node.  This does not accomplish that because:

- RPC requests are fast.  You *could* retrieve 1000 blocks quickly
enough from a high performance public RPC node to calculate the
checksum.
- Even if you couldn't, you could *easily* hack up oxend to proxy the
entire "test" request to some other node, thereby allowing you to run
multiple SNs without actually needing to store the blockchain.
- The testee gets I/O-trashed by having to go look up random blocks to
calculate the checksum for a test that really isn't useful.
- It is possible to abuse the blockchain test result feature by spamming
other storage servers to make other nodes waste significant resources to
compute these blockchain tests for random heights.
- The tests aren't actually *used* for anything: if you fail a test,
nothing happens.

Most importantly of all, this should never have been in storage server:
it has nothing at all to do with storing files, and is entirely outside
storage-server's purview to perform any such blockchain test: rather
that belongs in oxend (if it were to be performed at all).