Replace Oxen HTTP RPC with OMQ RPC
This replaces the HTTP RPC interface for communicating with the local
oxend with OxenMQ-based RPC requests.  This has advantages:

- Persistent, single connection rather than opening new connections
every time we need anything.
- We don't have to poll for new blocks every 1s: rather we can use
`sub.block` to have oxend push notifications to us when new blocks are
added; previously we had excessive (1s) polling which we needed to be
sure we noticed new blocks; now we avoid all of that *and* learn about
new blocks faster (since we get pushed notifications).
- Allows better host security (by protecting access to the unix socket
rather than needing a localhost port)
- Matches lokinet, which moved to use OMQ RPC in the last hard fork
(0.8).

Currently we still need HTTP JSON RPC requests to contact the bootstrap
nodes but that will eventually change as well.