Fix crypto_test
assert() is a no-op in non-debug builds, so the test was usually failing
to detect when things were broken.

This updates it to print stuff and do a proper run-time comparison.

In this case, the test was broken because it was calling encrypt_gcm
twice instead of encrypt_gcm+decrypt_gcm.  This fixes it.