Make integration test mode more obvious
Print a warning during startup, and handling the missing key options
gracefully (rather than dumping an exception to the overall main()
exception handler).

Print a critical error if the integration test options are missing or
invalid; without this an exception is thrown but it isn't obvious where
it comes from (especially if you forgot you compiled with integration
tests enabled).