Replace Monero crypto with libsodium
This replaces the Monero crypto implementation with a backwards
compatible libsodium implementation.  In particular, this replaces the
signature generation's random value with a deterministic value

    H(H(a) || A || M)

(reduced and clamped).  Since this deterministic value is perfectly good
as a "random" value, this avoids needing a RNG for generation, getting
us something closer (though still not) standard Ed25519 sigs.