Simplifies/unwraps the oxend_request endpoint
Redoes the oxend rpc API interface to make it slightly simpler and avoid
double-encoding (or triple-encoding!) json data.

The API now looks like:

    {"method": "oxend_request", "params": {"endpoint": "get_service_nodes", "params": {"limit": 1}}}

This renames the "oxend_params" key to just "params" (because the
"oxend" bit seems redundant given the method name) and makes it optional
(because many oxend rpc endpoints, including get_service_nodes, do not
require parameters to be passed at all).

The return value is now (when also using `"json": true` in the control
parameter added in the previous commit) straight JSON:

    {"status":200,"body":{"result":{"block_hash":"699e2f20bcb...

instead of:

    {"status":200,"body":"{\"result\":\"{\\\"block_hash\\\":\\\"699e2f20bcb...

If not requesting json embedding in the control parameters (see previous
commit) then there will still be the outer later of json string encoding
for body, but the JSON is still embedded directly rather than being an
extra string layer:

    {"status":200,"body":"{\"result\":{\"block_hash\":\"699e2f20bcb...

On error of whatever reason you get back a 400 status with the error
message in "body", for example:

    {"body":"Unable to parse request: Failed to parse JSON parameters","status":400}