Add direction to xchacha20 shared key
This alters the derived key hash to reverse the pubkey order when
replying so that the same key in each direction is used, i.e. for a
message from client C to snode S the key is:

    H(cS || C || S)

before this commit, the would use an encryption key for return messages
of:

    H(sC || S || C)

and, while the client can still decrypt that, it means the client has
*two* derived keys to worry about.  With this change, server swap the
order so it puts itself *second*:

    H(sC || C || S)

which will yield the same shared key as the client derived for the
original message.