Fix serialize on 32-bit arches
The size before the string was being serialized as a size_t, which
differs on different arches; fixed it to a uin64_t so that it matches
amd64.

Also remove an undesirable `reserve()`: reserving the amount needed for
an intermediate result tends to result in *more* allocations happening
rather than less.  Plus the reserve size was wrong (allocating only 4
bytes for the 8 byte size_t value).