Separate Database from boost::asio
Database has a dependency on boost::asio so that it can set up a timer,
but this is awkward as it couples the Database class with an
implementation detail of the Database user.

Fix this by removing it, making the cleanup timer callback the
responsibility of the caller.

This also fixes some spurious failures due to race conditions between
the threads in the storage test code.