Refactor how TTL & timestamps are stored
Currently we have an awkward storage of timestamp + ttl + expiry in the
database, and timestamp + ttl passed in from the client (as strings!).
Althis is awkward because we want to be able to shorten the expiry, but
that would mess up TTL.  Additionally everything is stored as
`uint64_t`s, which is messy and not type safe.

This commit makes these changes:

- ttl and timestamp can now be sent by the client as integers (in
addition to the current string value).  NB: this is not reliable until
the entire SN network is on the next SS release.

- ttl, timestamp, and expiry are now type-safe std::chrono types instead
of raw integers (milliseconds for ttl and system_clock::time_points for
the other two).

- ttl is no longer stored: instead we just store timestamp + expiry.
(This will let us update expiry later without worrying about TTL).

- ttl/timestamp value validation is moved out of `common/oxen_common.h`
(which was a very odd place for it) and into request_handler.

- serialization no longer supports message_t; rather message_t is *only*
for holding the value the client gives us.  SS now uses storage::Item
everywhere other than incoming client data.

- Removed TTL/Nonce storage and retrieval

- Fully specify the queries columns instead of using `SELECT *`

- Don't use "`" for identifier quoting inside sqlite.  It's non-standard
MySQL garbage that sqlite3 supported only for MySQL compatibility.