Support bt-encoded response (needed for forwarding)
When an incoming OMQ RPC storage.whatever client request is bt-encoded
we respond in kind.  This isn't all that critical for clients, but more
important this is required for bt-encoded comms between SNs, which is
very useful, particularly when synchronizing messages (bt-encodes bytes,
so is 75% of the size of base64 encoded values).

This required moving the json "dump()" call much closer to the response;
this lets us convert the json into bt_value in the protocol handler,
rather than pushing dumped json through the interface.

Aside from enabling bt-encoded response, this is also cleaner design.