Adds ed25519 pubkey to oxend pings
This will allow the oxend to verify the pubkey when pings are received
and reject the ping if they are wrong.

The reason this causes issues is that storage server only receives the
snode keys on startup and if oxend is restarted without restarting
storage server then it could end up running with the wrong keys.

If oxend instead rejects the pings and stops sending uptime proofs then
the operator will know that something needs attention and can restart
the storage server.