Fix broken nearest-swarm logic & optimizations
The nearest-swarm logic was rather broken when handling ids in the
wrapping space (i.e. > max swarm id or < min swarm id).  This fixes the
issues, as well as optimizing it.

Issues:
- all the wrapped-around values trying to compute distances were off by
  *2*:
  - The first off-by-one was because, even though swarm ids cannot be ==
    max-u64, swarm space values (from pubkeys) *can* be, so all the
    logic trying to make it wrap at MAX_SWARM_ID was invalid.
  - The wrapping at MAX_SWARM_ID was *also* wrong because, even ignoring
    the above point, distances are % (MAX_SWARM_ID+1), not %
    (MAX_SWARM_ID).
- Because of the above, a session ID that maps to 0xff..ff in swarm
  space would compute the distance to swarm_id 0 as max-u64 and thus
  always go to the top swarm_id, even when 0xff..fe would map to
  swarm_id 0.
- Values in the space between 0 and the minimum swarm id were rounding
  to the *right* side node, while everything else rounds to the left
  edge.

All of this is due to bad logic and insufficient test cases that
unfortunately was not properly reviewed when originally added to storage
server.

This commit changes the behaviour.  This means old and new nodes *will*
slightly disagree for a couple of edge cases:
- we *do* have a swarm 0 on mainnet, so a session id that maps to
  exactly 0xff..ff will break while the network has mixed versions.
  This, however, is extremely rare (~1 in 2^64), and so is unlikely to
  matter in practice.
- Values in the [0, lowest_swarm_id) that are exactly at the midpoint
  change to now prefer the left node (highest_swarm_id) rather than the
  right (lowest_swarm_id).  This makes it consistent with the other
  cases.  On mainnet, however, this case doesn't actually exist between
  lowest_swarm_id equals 0.

This also changes the code to use a binary search, rather than linear
scan, because swarms are guaranteed to be sorted.  This also
significantly simplifies the code.  The linear scan was just
unoptimized.