Merge namespace/data values into a single subscription
Maintaining multiple subscriptions for the same account on the same
connection for each different set of namespaces/want_data is painful, so
merge them into one.

This has some implications about allowing a single connection to renew
namespace and want_data subscriptions (even if those namespaces and/or
want_data flag aren't getting specified in renewals), but that is
relatively minor and it allows considerable simplification of
subscriptions.