Fix get_expiries when called with a single hash
get_expiries was not returning any expiries when called with a single
hash because of a bug in the database code's special-casing of a
single-hash-lookup that didn't bind the query parameters.

This fixes it.

As a temporary workaround until this makes it into a mandatory upgrade,
clients are suggested to include a second, fake hash `"fake"` when
requesting just one expiry, which will not return any results (because
it isn't an actual valid hash), but will avoid the bug and return the
desired message expiry in the response.