Make store and revoke update timestamps on conflict
When a store happens we de-duplicate the store (from the hash conflict),
but *don't* update the expiry timestamp.  This is undesireable (and a
bug): we want such stores to be able to de-duplicate but auto-extend the
existing message.  Without this, a store for a message that is just
about to expire could return true, but then the message would expire
anyway moments after the store.

The revoke_subaccounts change here is similar, though less important:
without it we would keep the last 50 revocations by insertion order,
*not* by the most recent revocation attempt, so if you have 50 revokes,
re-revoke the person at position 0, then add a new revocation, the
person at 0 would still get dropped.  With the change here, the
duplicate revocation would move them back to position 50, so the new
revocation afterward doesn't remove them.