Redo db.store() to better handle expiry updates
store()'s use of an upsert didn't allow us to tell what happened (i.e.
new row vs updated), but we need to know that because we only want to
send push notifications for new messages, but not expiry updates of
existing messages.

This thus moves away from an upsert and instead handles the queries we
need separately.  This might be a small performance improvement for
existing messages (because we now no longer have to provides the data to
sqlite at all for such a case), but more importantly it does what we
need.

The return type now changes as a result: we were using optional<bool> as
a tri-state, but we actually have *four* possible states:
- new message
- existing message that had its expiry update
- existing message that already had >= expiry
- full database
and these are now handled via an enum class return value.