Implement SQLite connection pool
The code previously shared an SQLite connection across threads using
sqlite's serialized mode, but that introduces race conditions: it's
entirely possible for weird interleaved results, or for
`sqlite3_changes` to return something for the wrong query.

This changes the internal database implementation to use a connection
pool instead so that we never use more than one connection concurrently:
we now maintain a stack of DatabaseImpls and pull one off the stack when
any code enters database operation space, then put it back on the stack
when done.

Initially we only have one connection on the stack (the initial one that
also does database initialization and an initial clean_expired call),
but new ones will be created on demand.  (As such the pool can never
shrink, but we do have limited concurrency that will prevent it from
growing indefinitely).

This also simplifies prepared statements: we no longer have to worry
about a lock and thread-specificity there since anything that can call
`prepared_st` is already pulled off the pool.  (There is a new lock in
pulling somethings off the pool, however, so this is probably net
neutral in terms of mutex locking).