Downgrade some warnings
These two cases *will* show up from time to time even when nothing is
wrong on the network (one when subaccount revocation is working as
expecting; and the other because Session doesn't always instantly know
about decomms and so might try an obsolete path).  As such, they
shouldn't be warning because they are neither a local nor a network
error and so shouldn't be warning.